Update strings.js
Removed the NLS support for this widget.  We do not have the resources to translate the strings.js files for  38 languages.  These files could be copied from Esri's client/stemapp/widgets/nearme widget and then modified to add the translations for the custom strings -- those beginning with "upstream" or "epa".